requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
requests>=2.31.0
pandas>=2.2.0
scikit-learn>=1.4.0
qdrant-client>=1.9.0
